Possibilités des noter avec des demies-étoiles

+++ db/migrate/20140513013145_create_ratings.rb
@@ -2,7 +2,7 @@ class CreateRatings < ActiveRecord::Migration
   def change
     create_table :ratings do |t|
       t.integer :album_id
-      t.decimal :rating
+      t.double precision :rating
       t.integer :user_id
 
       t.timestamps
